+++ CHANGELOG.md
@@ -29,6 +29,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Don't parse `#1` as a hashtag.
+- Added missing attributes to Mastodon API `Relationship` entity.
 
 ## [4.4.0] - 2025-06-05